Give a max width for gene panel tooltip
