+++ src/pages/studyView/TableUtils.tsx
@@ -4,6 +4,9 @@ import styles from "./table/tables.module.scss";
 import classnames from 'classnames';
 import DefaultTooltip from "../../shared/components/defaultTooltip/DefaultTooltip";
 import {ICON_FILTER_OFF, ICON_FILTER_ON} from "shared/lib/Colors";
+import {GenePanelList} from "pages/studyView/table/GenePanelModal";
+import {getFrequencyStr} from "pages/studyView/StudyViewUtils";
+import {GenePanel, GenePanelToGene} from "shared/api/generated/CBioPortalAPIInternal";
 
 export function getGeneColumnHeaderRender(cellMargin: number, headerName: string, cancerGeneListFilterEnabled: boolean, isFilteredByCancerGeneList: boolean, cancerGeneIconToggle: (event: any) => void) {
     return <div style={{marginLeft: cellMargin}} className={styles.displayFlex} data-test='gene-column-header'>
@@ -41,4 +44,29 @@ export function getCancerGeneToggledOverlay(cancerGeneFilterEnabled: boolean) {
 
 export function getCancerGeneFilterToggleIcon(isFilteredByCancerGeneList:boolean) {
     return <span data-test='cancer-gene-filter' className={classnames(styles.cancerGeneIcon, styles.displayFlex)} style={{color: isFilteredByCancerGeneList ? ICON_FILTER_ON : ICON_FILTER_OFF}}><i className='fa fa-filter'></i></span>;
+}
+
+export function getFreqColumnRender(numberOfSamplesProfiled: number, numberOfAlteredCases: number, matchingGenePanels: GenePanel[], toggleModal: (panelName: string, genes: GenePanelToGene[]) => void) {
+    const addTotalProfiledOverlay = () => (
+        <span style={{display: 'flex', flexDirection: 'column'}}>
+            <span>{`# of samples profiled for mutations in this gene: ${numberOfSamplesProfiled.toLocaleString()}`}</span>
+            <GenePanelList
+                genePanels={matchingGenePanels}
+                toggleModal={toggleModal}
+            />
+        </span>
+    );
+    return (
+        <DefaultTooltip
+            placement="right"
+            overlay={addTotalProfiledOverlay}
+            destroyTooltipOnHide={true}
+        >
+            <span>
+                {getFrequencyStr(
+                    (numberOfAlteredCases / numberOfSamplesProfiled) * 100
+                )}
+            </span>
+        </DefaultTooltip>
+    );
 }
